--- dotcom-rendering/src/lib/dynamicSlices.tsx
@@ -1,105 +1,0 @@
-import { LI } from '../components/Card/components/LI';
-import { UL } from '../components/Card/components/UL';
-import type { Loading } from '../components/CardPicture';
-import type { DCRContainerPalette, DCRFrontCard } from '../types/front';
-import { Card25Media25Tall, Card50Media50 } from './cardWrappers';
-
-/**
- * A place for shared slices which are used across multiple dynamic/ containers
- *
- * ASCII Art Guide:
- * '#'            => Image
- * ' ' | '_'      => text / trail / supporting content
- */
-
-/* ._________________.________.________.
- * |#################|########|########|
- * |                 |        |        |
- * |_________________|________|________|
- */
-export const Card50_Card25_Card25 = ({
-	cards,
-	containerPalette,
-	showAge,
-	absoluteServerTimes,
-	imageLoading,
-}: {
-	cards: DCRFrontCard[];
-	imageLoading: Loading;
-	containerPalette?: DCRContainerPalette;
-	showAge?: boolean;
-	absoluteServerTimes: boolean;
-}) => {
-	const card50 = cards.slice(0, 1);
-	const cards25 = cards.slice(1, 3);
-
-	return (
-		<UL direction="row" padBottom={true}>
-			{card50.map((trail) => (
-				<LI key={trail.url} percentage="50%" padSides={true}>
-					<Card50Media50
-						trail={trail}
-						containerPalette={containerPalette}
-						showAge={showAge}
-						absoluteServerTimes={absoluteServerTimes}
-						imageLoading={imageLoading}
-					/>
-				</LI>
-			))}
-
-			{cards25.map((trail) => (
-				<LI
-					key={trail.url}
-					percentage="25%"
-					padSides={true}
-					showDivider={true}
-				>
-					<Card25Media25Tall
-						trail={trail}
-						containerPalette={containerPalette}
-						showAge={showAge}
-						absoluteServerTimes={absoluteServerTimes}
-						imageLoading={imageLoading}
-					/>
-				</LI>
-			))}
-		</UL>
-	);
-};
-
-/**
- * Abstraction to decide whether to show padding on wrapped rows of cards.
- *
- * For three columns, We have different results with 5 or 9 cards
- *
- * @example - All but last 2
- * ```
- * ┌───┐ ┌───┐ ┌───┐
- * │Pad│ │Pad│ │Pad│
- * └───┘ └───┘ └───┘
- * ┌───┐ ┌───┐
- * │No!│ │No!│
- * └───┘ └───┘
- * ```
- * - All but last 3
- * ```
- * ┌───┐ ┌───┐ ┌───┐
- * │Pad│ │Pad│ │Pad│
- * └───┘ └───┘ └───┘
- * ┌───┐ ┌───┐ ┌───┐
- * │Pad│ │Pad│ │Pad│
- * └───┘ └───┘ └───┘
- * ┌───┐ ┌───┐ ┌───┐
- * │No!│ │No!│ │No!│
- * └───┘ └───┘ └───┘
- * ```
- *
- * @param index - Index of the current card
- * @param totalCards - Total number of cards being shown
- * @param cardsPerRow - No. of cards in each row (if full)
- */
-export const shouldPadWrappableRows = (
-	index: number,
-	totalCards: number,
-	cardsPerRow: number,
-): boolean => index < totalCards - (totalCards % cardsPerRow || cardsPerRow);
